Speech-to-Speech Conversational AI/requirements.txt
openai-whisper
gtts
gradio
groq
whisper
python-dotenv